+++ src/project/setup.ts
@@ -1,9 +1,9 @@
 import path from "node:path";
 import chalk from "chalk";
-import inquirer from "inquirer";
+import { addPackageDependencies } from "../helper/add-deps";
 import { templatesDir } from "../helper/constant";
-import { getAvailableTemplates } from "../helper/validator";
 import type { ProjectConfig } from "../types";
+import { getTemplatePath } from "../types/templates";
 import {
 	copyTemplate,
 	initGit,
@@ -29,8 +29,12 @@ export class ProjectSetup {
 	public async setupProject() {
 		try {
 			await this.setupBaseStructure();
+			await this.setupFrontendApps();
+			//TODO: implement this later
+			// await this.setupBackend();
 			await this.setupDatabaseIfNeeded();
-			await this.setupApps();
+			await this.setupAuthIfNeeded();
+			await this.setupAddons();
 			await this.initializeGitIfNeeded();
 			await this.installDependenciesIfNeeded();
 			this.displayNextSteps();
@@ -50,12 +54,117 @@ export class ProjectSetup {
 		console.log(chalk.green("✓ Base turbo-repo structure created"));
 	}
 
+	private async setupFrontendApps() {
+		if (
+			!this.projectConfig.frontend ||
+			this.projectConfig.frontend.length === 0
+		)
+			return;
+
+		console.log(chalk.yellow("\n🚀 Setting up frontend applications..."));
+
+		for (const app of this.projectConfig.frontend) {
+			console.log(chalk.yellow(`\n📱 Adding ${app} application...`));
+			const appPath = getTemplatePath({ frontend: [app] }, "app");
+			if (!appPath) continue;
+
+			const appTemplatePath = path.join(templatesDir, appPath);
+			const appTargetPath = path.join(this.projectPath, "apps", app);
+			await copyTemplate(appTemplatePath, appTargetPath, this.projectConfig);
+			console.log(chalk.green(`✓ ${app} application added`));
+		}
+	}
+
+	//TODO: implement this later
+	// private async setupBackend() {
+	// 	if (!this.projectConfig.backend || this.projectConfig.backend === "none") return;
+
+	// 	console.log(chalk.yellow(`\n🔌 Setting up ${this.projectConfig.backend} backend...`));
+	// 	const serverPath = getTemplatePath({ backend: this.projectConfig.backend }, "server");
+	// 	if (!serverPath) return;
+
+	// 	const serverTemplatePath = path.join(templatesDir, serverPath);
+	// 	const serverTargetPath = path.join(this.projectPath, "apps", "api");
+	// 	await copyTemplate(serverTemplatePath, serverTargetPath, this.projectConfig);
+	// 	console.log(chalk.green(`✓ ${this.projectConfig.backend} backend added`));
+	// }
+
 	private async setupDatabaseIfNeeded() {
-		if (!this.projectConfig.drizzle && !this.projectConfig.prisma) return;
+		if (
+			(!this.projectConfig.drizzle && !this.projectConfig.prisma) ||
+			this.projectConfig.database === "none"
+		)
+			return;
 
 		console.log(chalk.yellow("\n🗄️  Setting up database..."));
 		await setupDatabase(this.projectConfig, this.projectPath);
 		console.log(chalk.green("✓ Database setup complete"));
+	}
+
+	private async setupAuthIfNeeded() {
+		if (
+			!this.projectConfig.auth ||
+			this.projectConfig.auth === "none" ||
+			this.projectConfig.database === "none"
+		)
+			return;
+
+		console.log(chalk.yellow(`\n🔑 Setting up ${this.projectConfig.auth}...`));
+		const authPath = getTemplatePath(
+			{
+				auth: this.projectConfig.auth,
+				orm: this.projectConfig.drizzle ? "drizzle" : "prisma",
+				database: this.projectConfig.database,
+			},
+			"auth",
+		);
+
+		if (!authPath) return;
+
+		const authTemplatePath = path.join(templatesDir, authPath);
+
+		// if nextjs/t3 is selected, set up auth in the nextjs/t3 app directory
+		if (
+			this.projectConfig.frontend?.includes("nextjs") ||
+			this.projectConfig.frontend?.includes("t3")
+		) {
+			const appName = this.projectConfig.frontend?.includes("nextjs")
+				? "nextjs"
+				: "t3";
+			const appPath = path.join(this.projectPath, "apps", appName);
+			const authTargetPath = path.join(appPath, "server", "auth");
+
+			await copyTemplate(authTemplatePath, authTargetPath, this.projectConfig);
+			addPackageDependencies(appPath, {
+				dependencies: ["better-auth"],
+			});
+		} else {
+			//? idk if i even need to setup auth in the packages directory
+			return;
+			// const authTargetPath = path.join(this.projectPath, "packages", "auth");
+			// await copyTemplate(authTemplatePath, authTargetPath, this.projectConfig);
+		}
+
+		console.log(chalk.green(`✓ ${this.projectConfig.auth} setup complete`));
+	}
+
+	private async setupAddons() {
+		if (!this.projectConfig.addons || this.projectConfig.addons.length === 0)
+			return;
+
+		console.log(chalk.yellow("\n📦 Setting up addons..."));
+
+		for (const addon of this.projectConfig.addons) {
+			console.log(chalk.yellow(`\n🧩 Adding ${addon} package...`));
+			const addonTemplatePath = path.join(templatesDir, "packages", addon);
+			const addonTargetPath = path.join(this.projectPath, "packages", addon);
+			await copyTemplate(
+				addonTemplatePath,
+				addonTargetPath,
+				this.projectConfig,
+			);
+			console.log(chalk.green(`✓ ${addon} package added`));
+		}
 	}
 
 	private async initializeGitIfNeeded() {
@@ -80,75 +189,6 @@ export class ProjectSetup {
 		await installDependencies(this.projectPath, this.projectConfig);
 		console.log(chalk.green("✓ Dependencies installed"));
 	}
-
-	private async setupApps() {
-		const appsPath = path.join(templatesDir, "apps");
-		const availableApps = await getAvailableTemplates(appsPath);
-
-		if (availableApps.length > 0) {
-			const { addApps } = await inquirer.prompt([
-				{
-					type: "confirm",
-					name: "addApps",
-					message: "🚀 Would you like to add any apps to your workspace?",
-					default: true,
-				},
-			]);
-
-			if (addApps) {
-				//TODO: what if user wants to add multiple apps?
-				const { app } = await inquirer.prompt([
-					{
-						type: "list",
-						name: "app",
-						message: "Select apps to add:",
-						choices: availableApps,
-					},
-				]);
-				console.log(chalk.yellow(`\n📦 Adding app: ${app}`));
-				const appTemplatePath = path.join(appsPath, app);
-				const appTargetPath = path.join(this.projectPath, "apps", app);
-				await copyTemplate(appTemplatePath, appTargetPath, this.projectConfig);
-				console.log(chalk.green(`✓ App ${app} added`));
-			}
-		}
-	}
-
-	//TODO: implement this
-	// private async setupPackages() {
-	//     const packagesPath = path.join(templatesDir, "packages");
-	//     const availablePackages = await getAvailableTemplates(packagesPath);
-
-	//     if (availablePackages.length > 0) {
-	//         const { addPackages } = await inquirer.prompt([
-	//             {
-	//                 type: 'confirm',
-	//                 name: 'addPackages',
-	//                 message: '📦 Would you like to add any packages to your workspace?',
-	//                 default: true
-	//             },
-	//         ]);
-
-	//         if (addPackages) {
-	//             const { selectedPackages } = await inquirer.prompt([
-	//                 {
-	//                     type: 'checkbox',
-	//                     name: 'selectedPackages',
-	//                     message: 'Select packages to add:',
-	//                     choices: availablePackages.filter(p => p !== 'database')
-	//                 },
-	//             ]);
-
-	//             for (const pkg of selectedPackages) {
-	//                 console.log(chalk.yellow(`\n📦 Adding package: ${pkg}`));
-	//                 const pkgTemplatePath = path.join(packagesPath, pkg);
-	//                 const pkgTargetPath = path.join(this.projectPath, "packages", pkg);
-	//                 await copyTemplate(pkgTemplatePath, pkgTargetPath, this.projectConfig);
-	//                 console.log(chalk.green(`✓ Package ${pkg} added`));
-	//             }
-	//         }
-	//     }
-	// }
 
 	private displayNextSteps() {
 		const packageManager = this.projectConfig.pnpm